# PERF: make Styler default formatter arguments statics instead of repeated dynamic